BugFix: attempt to solve copy selection problem (#3854)
* BugFix: attempt to solve copy selection problem

We have established that the Copy as HTML and Copy as Image modes do not
seem to have the problem that the plain Copy (as Text) one has. It is
noticeable that the first two have the same process to select the portion
of the `TBuffer` that is to be copied and that seems to be different to
that of the third. So this changes the latter to use the same system that
has been refactored out of the other two.

Also, since all the text we need is already in an arrangement that we can
actually copy quite fast (a `QStringList`) there is no need to do a slow
`QChar` by `QChar` copy of every `QChar` in the whole selection - instead
we can take a copy of that section of the `TBuffer::lineBuffer` that
contains the wanted lines - and then just trim off the leading `QChar`s
on the first line that precede the selection stop (and replace them with
spaces if it is a multi-line selection) and chop off the unwanted trailing
`QChar`s on the last line... This way of doing things also avoids the
complication that empty intermediate lines might not have ANY characters in
them!

Also - I noted that the code in the Copy as HTML and Copy as Image that
was similar to the code in the existing `normalisedSelection()` was
actually faulty as when the start and end points were on the same line
(i.e. the Y values were the same), the test to see whether the points had
to be swapped had an incorrect test for the X values also being the same,
whereas they should have been the end point X value being less than the
starting X value - as they are in the normalising method!

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>

* BugFix: get select all working in TConsole/TTextEdit

It needed a `normalisedSelection()` call to transfer the values manually
assigned to `(QPoint) mDragStart` and `mDragSelectionEnd` to the `mPA` and
`mPB` ones that are used when iterating through the selection.

This then showed that it was necessary to force an update of BOTH
`TTextEdit` instances for a `TConsole` so that a selection made in one
was shown in both if they both contain selected text - which is bound to
happen if ALL the text is selected.

A further click should *clear* the selection of ALL the text and start
a new drag operation.  This is now more likely to happen, but it does not
seem totally totally correct to me yet.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>