Refactor: remove dead code for mLF_on_GA (#3827)
Although the `(bool) cTelnet::mLF_ON_GA` is initialised to `false` it is
always set to `true` by the same variable in the `Host` class - which is
never unset. Therefore the execution where this is used as a condition is
fixed to that for a `true` state. Thus it can be eliminated and the code
cleaned up from it's removal.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>