Enhance: make compact input line act on all profiles
Makes the compact input line a toggle (so now it also has constant text)
and act on all profiles.  The previous code only considered the currently
active profile which is contradicted by the fact that the state - which was
and is maintained between sessions - is stored centrally as an application
setting rather than a per profile one.

This was split out of a larger PR that originally included it:
https://github.com/Mudlet/Mudlet/pull/3844 .

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>