Revise: improve Variables icon
This adds a white outline around the otherwise entirely black icon so that
it becomes visible when used with a dark desktop theme.

For the record this icon was created from scratch on a 256x256 transparent
background using The G.I.M.P. to draw "X=" in text in black using the
"Serif" font at a size of 150, that image was then redrawn in white four
times and each of them was offset by +/-4 in complimentary X and Y
directions so that they expanded the background white around the black
foreground.

The reason for expanding from the original 48x48 pixel size was to make the
anti-aliasing take up a much smaller proportion of the image - so it will
appear sharper even at higher resolutions.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>