Revise: refactor stuff in preparation to add fixed strings of name/password
I can see that there are some other comboBoxes on the preferences that it
would be useful to have repopulated immediately that the language is
changed alongside the new one I want to put in to support the user being
able to choose a fixed string that has the password and probably the
character name in:

* convert single row/column QGridLayouts to corresponding Q(V|H)BoxLayouts
in the profile preferences form - they automatically keep their elements
in order compared to the GridLayout form which randomises them if any are
edited.

* add buddies to those controls with associated labels which are not,
actually associated in that way.

* rename `(QListWidgeti) dlgProfilePreferences::dictList` to
`dlgProfilePreferences::listWidget_dictionaries`

* rename `(QComboBox) dlgProfilePreferences::comboBox` to
`dlgProfilePreferences::comboBox_mapBackups`

* refactor code that populates `dlgProfilePreferences::listWidget_dictionaries`
to a separate:
`dlgProfilePrefences::generateDictionaryTexts()`

* refactor code that populates `dlgProfilePreferences::comboBox_encoding`
to a separate:
`dlgProfilePreferences::generateEncodingTexts()`

* add `generateDictionaryTexts()` and `generateEncodingTexts()` to to:
`slot_guiLanguageChanged(...)`

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>