Enhance: add capability to have custom login texts involving password
This currently only offers one text - of the form:
"connect {character name} {password}"
and can be selected from a new comboBox on the "Special options" tab of the
preferences. It is saved with the profile data.

It is also subject to the same timeout of the `sendCharacterPassword()`
function.

The `doLogin()` function in the external `LuaGlobal.lua` file has been
modified to check whether the option to use a custom login text has been
set and will send the chosen string two seconds after a successful
connection instead of the character name and password after two and three
seconds respectively.

Obviously if a more sophisticated system is required the user could script
it or a package/module supplied by the MUD could be provided. The custom
strings could also be appended to if a compelling case is made to the
Mudlet Makers though for backwards compatibility it would not be
recommended to remove a string previously added.

For user display purposes the string shown in the preferences is expected
to be translated for the tokens "{character name}" and "{password}" however
the system is designed to not have the actual string changed on a per
locale basis.

Also, whilst editing the profile preferences dialogue/form I found some
widgets were not include in or were totally out of tab-order sequence; so
I have revised that to be more natural. In doing so I also felt that the
"Special options" tab should be re-positioned to be the last tab - so it
and the "connection" one have swapped places (and the tab order adjusted
as appropriate.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>