Revise: reorder tabs and tab stops in profile preferences
Switch some uses of QGridLayout <==> QVBoxLayout depending on whether there
is more or only one column of items.

Moved the "Special options" tab to the last position.

Switch some single columns of items to become two columns so that the
overall vertical size needed for the dialogue can be reduced.

Moved the Discord and IRC options to a new tab "Chat".

These last two are important as they allow the overall dialogue to be
smaller and thus may enable us to close: #3467, #3304.

Removed unused groupBox that has something about deleting map backups. It
is not handled or mentioned in the current code at all and just sits there
as a disabled widget in the mapper tab.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>