Enhance: add functions for handling game character details
`getPlayerName` takes no arguments and returns the "Character name" from
the connection screen. There is no provision to change this stored setting.

`sendPlayerName` takes no arguments and causes the "Character name" from
the connection screen to be sent without echoing to the game server, it
will return a nil plus an error message should the name not be set (is
empty string).

`sendPlayerPassword` takes no arguments and causes the "Password" from
the connection screen to be sent without echoing to the game server, it
will return a nil plus an error message should no password be set (is empty
string). There is no provision to find out this stored setting.

`sendCustomLoginText` takes no arguments and causes the custom login text
to be sent provided it is not disabled (set to the 0th one).

`getCustomLoginTextId` gets the index (from zero upwards) of the custom
login text - so at this point will return `0` {disabled} or `1`.

Remove saving and restoring login name and password either side of
package/module loading - we do not need them to protect the details from
being changed in a loaded XML file because those pieces of information are
not stored in there.

Move the existing timed sending of the user name and password to a new
`doLogin()` Lua function in the external `LuaGlobal.lua` file which
effectively does the same thing but can be replaced by something more
sophisticated if the user needs it. It also checks whether the option to
use a custom login text has been set and will send the chosen string two
seconds after a successful connection instead of the character name and
password after two and three seconds respectively.

Alternatively the user can replace it with a dummy (empty) function with
the same name and use a pair of triggers (which I have tested) or an
`sysConnectionEvent` event handler.

The custom login currently only offers one text - of the form:
"connect {character name} {password}"
which can be selected from a new comboBox on the "Special options" tab of
the preferences. It is saved with the profile data.

The custom strings could also be appended to if a compelling case is made
to the Mudlet Makers though for backwards compatibility it would not be
recommended to remove a string previously added.

For user display purposes the string shown in the preferences is expected
to be translated for the tokens "{character name}" and "{password}" however
the system is designed to not have the actual string changed on a per
locale basis. The UI texts are generated by a method that detects whether
it is being used for the first time on dialogue creation (and so considers
the existing profile's setting) or after a GUI language change - which
keeps the same setting but shows it in the new language. It is my intend
to refactor out the population of some of the other combo-boxes in that
dialogue as well so that the user will not have to restart Mudlet just to
get other important settings in a language that they can comprehend.

Also:
* demote dlgConnectionProfiles::slot_connectToServer() from a slot to a
normal method - as it is not connected up to the signal/slot system.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>