bugfix: Point map config alias to correct help
making the #3970 fix in the script broke `map help config`. because the `show_help` function parses the user supplied argument, the script's so-called "help file" needs to have the same singular title as the requested argument. thus, this commit points `map config` to `map.help.config` when the alias is called without argument and also retains the intended `map help config` feature.