Room name display and font sizing fixes
Room name display is now independent of room number display.
The cut-off for the font size is lower (4 instead of 7).

Room names are scaled as if the number 88 would be printed in the room
box, which works for all practical purposes.

Font size discovery is now done in floating point, and by scaling with
factors 1.2 (up) and 1.05 (down) which is faster *and* looks way better.