Add a separate GET operation
Sometimes you just need to read some data from the HTTP server without
going through writing the response to a file.

Also, when a POST redirects to a GET, the reply was lost.

Thus this change introduces a dedicated "getHTTP" function, and a
"sysGetHttpDone" event that contains the actual data.