Emit the debug console's output to standard error
The debug console only has a couple of lines, and scrolling it back is
painful, esp. if you want to figure out why your script runs into an
endless loop.

This simple one-line patch also emits the debug console's content to
standard error which can be captured in a file or in the system
terminal's scrollback buffer.