BugFix: restore L. yellow colour button + insert new control in tab order
Due to a numbering error the button used to set the ANSI colour for Light
Yellow for the profile's console was positioned to be on top of the Light
Green one. This was introduced by my #3929 a few days ago!

Also insert a recently added checkbox (force Telnet CHARSET negotiation
off) did not have a tab order specified so navigating around the
preferences dialog with the tab key would not visit it in the right
sequence.

For consistency the order of the main console font size and anti-aliasing
checkbox needed to be swapped over.

To allow both PRs #3952 and #3832 to merge in where they both add in a new
(different) control into the `gridLayout_groupBox_debug` layout in the
"Special options" tab of the profile preferences dialogue a gap has
been left in row 3 of that layout which will be used by the latter of those
two PRs (the other one will add a new row to the end).

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>