Bugfix: handling of Big5 encodings for ambiguous E Asian auto-widening
When the preference that controlled the E. Asian widening was set to "Auto"
it was not being correctly detected as the wrong `QByteArray` strings were
being checked against the currently selected Game Server encoding setting.
This would mean that both the variants that Mudlet offered would be using
the "Normal"/"Narrow" wide for such ambiguous graphemes rather than the
"Wide" that was intended.

Also, there were some places in the `TLuaInterpreter` class where the
encoding details were being treated as `QString`s rather than the
`QByteArray` that is used specifically to match up with Qt's own usage,
and avoid the whole `QStringLiteral(...)`/`QObject::tr(...)` wrapper issue
that arises from the use of `QString`...

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>