Enhance: add Lua API for map symbol font
Adds `mapSymbolFontInfo((string) fontName[,
                        (bool) onlyUseThisFont[,
                        (float) scaling factor]])`

and `mapSymbolFontInfo()`.

The former takes:
* a font name (default is currently: "Bitstream Vera Sans Mono" which can
be also be specified with an EMPTY {i.e. `""`} string. Alternatively,
should the other arguments want to be set but not this one a `nil` is also
acceptable for this argument to not change the font.
* whether to only use the specified font or to use any that is available,
use `true` or `false` to force the setting, or if it is `nil` (or not
provided) no change is made so that the third argument can still be
provided.
* the scaling factor (between 0.50 to 2.00) that can be used to tweak
the scale (default is 1.00) compared to the calculated font size that
should enable an 'M' to fit inside the room rectangle or circle.

The latter returns a table with the following elements:
* (string) the (family) "fontName"
* (bool) "onlyUseThisFont"
* (float) "scalingFactor"

Note that both of this functions operate only when a map is loaded as the
map symbol font details are store in the map file (and NOT in the profile).

This PR should close #4031.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>