BugFix: handle floating toolbars with no DockArea set
I discovered I had a problem with this when working on something else with
an existing profile. The result was that the toolbar concerned was not
added to the main window. This PR forces a resolution to this by reapplying
the toolbar to the default left DockWidgetArea if no previous one seemed to
be set.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>