BugFix: allow color changes mid-MBCS (part 2)
This commit is about allowing the fore/back ground colors to be different
in the left and right halves of a grapheme - by enabling a second format
to be stored as part of a single `TChar` instance - and handling the
additions needed to display this.

It adds new Lua API functions
* setSplitFgColor([windowName ,] red, green, blue)
* setSplitBgColor([windowName ,] red, green, blue)
* resetSplitFormat([windowName])
which act on the FIRST character in the current selection to change the
colours used for the LEFT side of that grapheme. This is so that the
majority of the existing codebase can be agnostic to the extra colours
on such characters - only the `getTextFormat()` function will return
extra data in a 'splitFormat' sub-table, namely extra `foreground` and
`background` colour component tables.

Note: This only half of the solution to #3939 the other half is to parse
the colour codes from between the two bytes that make up a DBCS encoded
grapheme in a (usually) Big5 encoded MUD/MOO server output data stream.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>