BugFix: repair failure to load Lua modules from profile directory
This was introduced by PR #3889 which was using a function defined in the
LuaGlobal.lua file before it was loaded! The function definition has been
moved to be within the TLuaInterpreter class C++ code and, as it is now
present in two places, refactored out to a helper method so it can be
loaded into both the main and the lua code formatter Lua interpreters.

This should close #4047.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>