Enhance: setup CI building for Mingw-w64 (part 1)
This adds three shell scripts that try to replace some powershell ones to
produce both 32 and 64-bit Windows packages. At present they should each
produce a zip archive file containing all the files needed to execute the
Mudlet application on a Windows PC that does NOT have MSYS2 installed in
it. Further work is needed to get excecutable installers built from that
set of files and to upload them to the *normal* location - at present they
will be available as build artifacts on the AppVeyor CI platform...

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>