BugFix: move to package directory BEFORE zipping things up
... otherwise we archive the whole project files, source code, git repo
and everything else. This was caused by move the archive file creation to
the new appveyor.after_success.sh script, it also means that it is no
longer possible to capture that archive and store it internally to AppVeyor
as a build artifact - so that bit needs to come out of the YAML control
file.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>