BugFix: try and prevent crashing during highlighting (#4005)
It seems that one of the tests in `(void) TTextEdit::highlightSelection()`
was being done too late - with the result that out of range indexing was
being attempted on the `(std::deque<std::deque<TChar>) TBuffer::buffer`
member. Specifically the test for `currentY >= totaly` should have been
done before `currentY` was used to select one from the outer `std::deque`
container so as to establish the size of it (to be stored as `maxX`).

The revised code replaces both the highlight and un-highlight code nested
loops with components that are more conventionally used, i.e. with the
test being based on a less than condition only with the limit being
determined within the top parts of the for loops and no additional tests
to break from the loop within the body thereof...

 BugFix: fix off-by-one errors in TTextEdit::getSelectedText(...)

The previous code did not allow for the fact that the actual end of line
and the end of the lines numbers are one less that the
`std::deque<T>::size()` method returns.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>
Co-authored-by: Manuel Wegmann <60551052+Edru2@users.noreply.github.com>
Co-authored-by: Vadim Peretokin <vperetokin@gmail.com>