Revise: make debug output less spammy for unprintable codepoints (#3832)
This is a per profile option, saved with the profile, that disables the
reporting of "Unprintable" codepoints in a `TConsole` (or rather the upper
pane - i.e. a `TTextEdit` thereof). Instead, in all cases, a count of all
the problematic codepoints are printed just before the TTextEdit
is destroyed. The previous behaviour - to report every single problem as it
occurs is now enabled with a new checkbox in the Special options tab of the
preferences.

Format the U+xxxx codepoint numbers in the report messages to be 4 hex
digits long as a minimum.

I thought at one point that the replacement character glyph is typically
2 spaces (a "wide" character) wide but in testing I found that not to be
the case.

For some reason the MacOs Clang builds on GH Actions and Travis were
failing on providing a `std::tuple<uint, std::string>` element when it
was provided as, say, `std::tuple{1, "Unassigned"}` so this commit tries to
help by using `std::tuple{1, std::string{"Unassigned"}}` to construct the
second element.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>