Horizontal scrollbars for all consoles + scrollbar fixes (#4085)
* fix scrollbar and allow horizontal wheel

* fix for characters didn't get cleared

* enable/disable horizontal scrollbar

* fix issues and resize console if horizontal scrollbar show/hide

* get scrollbarheight dynamically

* change layout design

make horizontal scrollbars part of the main console. which means they get also resizes if borders resize. This looks much better then before and the code for the resize event is not needed anymore.

* remove unneeded console resize

* update scrollbar when lowerPane hides due to scrolling

and also if it hides due to control+return

* add horizontal scrollbar to geyser miniconsole by cons

* fix issues #3054

fix #3054 and alll issues described in it

* Update TCommandLine.cpp

forgot one

* fix changing size

fix changing size and got rid of bufferscrollup as it wasn't needed and didn't do what it was supposed to anyway

* get rid of forgotten debugging messages

* scroll horizontal by highlighting text

horizontal scroll by highlighting the same way vertical scroll does

* horizontal scrolling was to fast + reset position

scrolling for horizontal seem to be to fast if the same values as for vertical are used.

horizontal position gets resetet if new lines appear the same way it happens for vertical scrollbar (it doesn't happen for if the split is active)

* set mMaxHRange to 0 if to small

* copy/paste fail spottet by @atari2600tim

e->delta() should be deltaY

* suggestions