Testing: a check to see that the AppVeyor CI can still build a snapshot
There are some revisions and some test behaviour but this is not complete
enough for anything other than PR snapshot builds. It aims to test to
see if messages can be sent to AppVeyor message listing and build artifacts
produced without listing them explicitly in the yaml file.

It also merges in recent commits in the development branch.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>