db:create shared databases
Allows for a 3rd parameter on db:create that is a boolean. If true, it will create the database outside of the profile at Mudlet/databases. If missing, previous functionality works the same. I will be testing this locally to see what kind of impact shared databases cause, but I believe folks are already doing through this profile-sharing.