BugFix: fix a CI script error around date calculations
The MacOs build was producing an error message:
CI/travis.osx.after_success.sh: line 60: [[: %cs: syntax error: operand expected (error token is "%cs")
== Creating a public test build ==
Deploying ./Mudlet PTB.app

It turns out that when using the `git show` command it is helpful to place
double quotes around the format specifier - especially (but not in this
case if there was spaces in the text) it isn't clear why the `%cs` was not
being handled but the fact that it was mentioned in the line where the
date number that it was suppose to produce was being used means that
it was not being handled by the original `git show`. Anyhow, for safety's
sake I have quoted it in all three CI scripts where it is used.

At the same time I have also renamed the variables that it was being stored
in and tested against to the SCREEMING_SNAKE_CASE form, as that form makes
bash variables a bit easier to pick out in scripts.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>