Revise: partial revisions but NOT ready to go yet [skip gha][skip travis]
The snapshot build code should now be complete. The PT builds should also
work now. However the release build is still not complete. For testing
purposes the NEXT commit should try and use SCP to upload the snapshot
code so that it can be verified to be usable for the release builds - this
needs some extra encryption stuff and a key (and salt) will probably need
to be created to use the MSYS2 `scp` rather than the WINSCP NET module
that the original PowerShell implimentation used.

Includes a debug line to locate the secure-file utility that might be
needed.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>