Enhance: improve PTB differentiation
This PR is dependent on #4089 (to provide the `./icon/mudlet.ico` Windows
Icon file).

It adds missing support to the CMake build process for the different
Windows application icons and provides a PTB variant for the installation
screen on Windows (currently the standard Mudlet release version is used
which is a little confusing!). It also adds the Windows PTB `.ico` to the
installer executable (a renamed setup.exe).

It also updates the qmake project file to include the six MacOs and Windows
icon files (`.icns` and `.ico` respectively) in the files that Qt Creator
will be aware of.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>