Revise: take out unneeded uses of cygpath [skip gha][skip travis]
When using native (to MSYS2) paths (with '/' directory separators) it is
probably not necessary to use the $APPVAYOR_BUILD_DIR - we can hard code
the path '/c/projects/mudlet/` which makes the quoting and processing of
paths simpler - if harder to revise later one if that value ever changed.

Also needed to fix some uses of `rsync` with a wild-card, especially when
the path containing them is double-quoted. If rsync is processing the
wildcard (instead of the shell?) it will stop at directories if a single
`*` is used - apparently one needs to use s `**` to cause it to decend.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>