Clean up console access code (#4186)
* Add a new Host::findConsole function

* Sketch: TLuaInterpreter.cpp modification

* Fix macros to eval their argument once

* More adaptions

* TLuaInterpreter: use WINDOW_NAME macro.

Some cases not touched because difficult.

* macro-ify a lot of functions

* Remove no-longer-needed accessors

* Clean up font and console handling

* unify setFontSize and setMiniConsoleFontSize
  the latter is now an alias to the former
* TConsole methods don't need "MiniConsole" in their name
  that's redundant
* teach Host::updateConsolesFont() to refresh the main console
  no need to expose these details

* Fix font setting

* Removed a bunch more unused methods from mudlet.cpp

* replace HostManager::getHostList

It was only used to
* create a list to test membership against
  just call getHost and check for null
* create a list of host names, each of which is then looked up
  which is unnecessary work

* Remove the "QMap<Host*, T*>" anti-pattern

It makes much more sense to add a T* member to the Host class.

* Allow window names beyond the Lua stack

simply treat a missing parameter (or nil) as empty string

* Use a nested class for iteration

* Fix host iterator

* move getLines() to Host

* remove mudlet::mTabMap

Redundant, the host map and Host::mpConsole works.

* more moves to host

* Move host-close code to the host

* Move even more stuff from mudlet:: to THost::

Also drop mudlet::mpProfilePreferencesDlgMap in favor of a pointer in
THost. Sorry that this got mashed in but I was halfway through this
before I noticed and didn't want to do the work twice.

* These things are pointers, thus should be named as such

as annoying as that is …

* Vanquish another two pointer-to-host maps

* Move host-specific layout updates to THost

This patch also removes layering violations in TToolbar and TDockWidget.

* mudlet.h include directives cleanup

* appease CodeFactor

* variable rename

* Check for valid dialog

* minor cleanup

* Revert mistaken checkin of test code

from d918164c77382332ad1cd9bd31b2d1ea173267fd

* rename function in error messages

* Crash fix

* renamed args for clarity

* Braces please

* Bugfix: "echo" didn't talk to labels

* Found two wrongly-named exceptions

* Clean up merge differences

* Merge error

* Fix potential NULL deref

* Re-add the option to have a prefs dialog without a host

* Don't delete the prefs dialog on close

* Fix "Don't delete the prefs dialog on close" another way.

Apparently Qt correctly handles this when using a QScopedPointer but not
using a QSharedPointer. That is, even though the shared pointer still
references the dialog, it's killed off without clearing the pointer.

Since keeping the dialog around forever isn't a good idea, replace with
a QPointer which *is* cleared.

Co-authored-by: Vadim Peretokin <vperetokin@gmail.com>